+++ NEWS.md
@@ -25,6 +25,7 @@ New features
 
   - New configuration ``dice_d4`` which provide four-sided dice in six color schemes (color controlled by their suit).
     Tetrahedrons with the rank as a numeral at the top point (#286).
+  - New configuration ``dice_numeral`` which provide six-sided dice with numerals instead of pips in six color schemes (#311).
   - New configurations ``dominoes_chinese`` and ``dominoes_chinese_black`` which provide [Chinese dominoes](https://en.wikipedia.org/wiki/Chinese_dominoes) and [Asian-style dice](https://en.wikipedia.org/wiki/Dice#Arrangement).
   ``dominoes_chinese`` has a white background with black and red pips while ``dominoes_chinese_black`` has a black background with white and red pips (#308).